use int for opening fd

+++ dttools/src/copy_stream.c
@@ -77,7 +77,7 @@ int64_t copy_fd_to_fd(int in, int out)
 
 int64_t copy_file_to_file(const char *input, const char *output)
 {
-	int64_t in = open(input, O_RDONLY);
+	int in = open(input, O_RDONLY);
 	if (in == -1)
 		return -1;
 	struct stat info;
@@ -86,7 +86,7 @@ int64_t copy_file_to_file(const char *input, const char *output)
 		return -1;
 	}
 
-	int64_t out = open(output, O_WRONLY | O_CREAT | O_TRUNC, info.st_mode & (S_IRWXU | S_IRWXG | S_IRWXO));
+	int out = open(output, O_WRONLY | O_CREAT | O_TRUNC, info.st_mode & (S_IRWXU | S_IRWXG | S_IRWXO));
 	if (out == -1 && errno == ENOTDIR) {
 		char dir[PATH_MAX];
 		path_dirname(output, dir);
